+++ app/routes/components/Switch/index.tsx
@@ -4,8 +4,8 @@ import { Section } from "src/components/Section";
 import type { Route } from "./+types/index";
 
 import { changelog } from "./changelog.codegen";
-import { Demo } from "./Demo";
-import { demoSourceCode } from "./Demo_source_code.codegen";
+import Demo01 from "./demos/Demo01";
+import demo01SourceCode from "./demos/Demo01.source_code.codegen";
 import { sourceCode } from "./source_code.codegen";
 
 export function meta(_: Route.MetaArgs) {
@@ -18,12 +18,11 @@ export function meta(_: Route.MetaArgs) {
 export default function SwitchPage() {
   return (
     <ComponentTemplate component="Switch">
-      <Section
-        title="Switch"
-        demoAndCode={[<Demo />, demoSourceCode]}
-        sourceCode={sourceCode}
-        changelog={changelog}
-      />
+      <Section.Root title="Switch">
+        <Section.Demo node={<Demo01 />} code={demo01SourceCode} />
+        <Section.SourceCode files={sourceCode} />
+        <Section.Changelog changelog={changelog} />
+      </Section.Root>
     </ComponentTemplate>
   );
 }