+++ src/components/Switch/index.ts
@@ -1,3 +1,3 @@
-// Switch from copy-ui @ 2025-10-02.
+// Switch from copy-ui @ 2025-10-14.
 
 export { Switch } from "./Switch";